requirements.txt
absl-py==2.1.0
aenum==3.1.15
alabaster==0.7.13
astor==0.8.1
attrs==23.2.0
babel==2.14.0
backcall==0.2.0
bddl==1.0.1
charset-normalizer==3.3.2
cloudpickle==2.2.1
commonmark==0.9.1
cycler==0.11.0
cython==0.29.1
decorator==5.1.1
docutils==0.19
exceptiongroup==1.2.1
fastjsonschema==2.20.0
ffmpeg-python==0.2.0
fonttools==4.38.0
freetype-py==2.4.0
future==1.0.0
gast==0.2.2
gin-config==0.1.3
gitdb==4.0.11
gitpython==3.1.43
google-pasta==0.2.0
gputil==1.4.0
grpcio==1.62.2
gym==0.26.2
gym-notices==0.0.8
h5py==3.8.0
idna==3.7
imageio==2.2.0
imageio-ffmpeg==0.2.0
imagesize==1.4.1
importlib-metadata==6.7.0
importlib-resources==5.12.0
iniconfig==2.0.0
ipython==7.34.0
jedi==0.19.1
jinja2==3.1.4
jsonschema==4.17.3
jupyter-core==4.12.0
jupytext==1.15.2
keras-applications==1.0.8
keras-preprocessing==1.1.2
kiwisolver==1.4.5
markdown==3.4.4
markdown-it-py==2.2.0
markupsafe==2.1.5
matplotlib==3.5.3
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.5
mdurl==0.1.2
nbformat==5.8.0
networkx==2.6.3
numpy==1.21.6
opencv-python==4.10.0.82
opt-einsum==3.3.0
packaging==24.0
pandas==1.3.5
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==9.5.0
pkgutil-resolve-name==1.3.10
pluggy==1.2.0
progressbar==2.5
prompt-toolkit==3.0.45
protobuf==3.20.0
psutil==6.0.0
ptyprocess==0.7.0
py360convert==0.1.0
pybullet-svl==3.1.6.4
pygments==2.17.2
pyparsing==3.1.2
pyrsistent==0.19.3
pytest==7.4.4
python-dateutil==2.9.0.post0
pytz==2024.1
pyyaml==6.0.1
recommonmark==0.7.1
requests==2.31.0
scipy==1.7.3
six==1.16.0
smmap==5.0.1
snowballstemmer==2.2.0
sphinx==5.3.0
sphinx-markdown-tables==0.0.17
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==2.3.0
toml==0.10.2
tomli==2.0.1
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.66.4
traitlets==5.9.0
transforms3d==0.4.1
trimesh==4.4.0
typing-extensions==4.7.1
urllib3==2.0.7
wcwidth==0.2.13
werkzeug==2.2.3
wrapt==1.16.0
wheel
yacs==0.1.8
zipp==3.15.0
